Fix app.kubernetes.io labels to use $ instead of . in vaultwarden chart (#77)
This commit changes the way the `app.kubernetes.io/name` and `app.kubernetes.io/instance` labels are generated in the vaultwarden Helm chart. Previously, the `.Values` object was used directly within the template expression, which is not allowed by Helm. This change updates the expression to use the `$` symbol, which correctly references the current values context.